refactor : 전체 패키지 위치 변경 및 조정
auth 패키지를 생성하여 로그인, 로그아웃, 계정 생성과 이메일 인증에 관한 도메인을 이동
