+++ src/main/java/inu/spp/cryptocoffee/auth/user/controller/ReissueController.java
@@ -1,0 +1,25 @@
+package inu.spp.cryptocoffee.auth.user.controller;
+
+import inu.spp.cryptocoffee.auth.user.service.ReissueService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@Tag(name = "Token Reissue API")
+@RestController
+public class ReissueController {
+
+    private final ReissueService reissueService;
+
+    @Operation(summary = "토큰 재발급")
+    @PostMapping("/api/users/reissue")
+    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) {
+        return reissueService.reissueRefreshToken(request, response);
+    }
+}
