+++ src/main/java/inu/spp/cryptocoffee/auth/user/dto/UserRoleEnum.java
@@ -1,0 +1,6 @@
+package inu.spp.cryptocoffee.auth.user.dto;
+
+public enum UserRoleEnum {
+    ROLE_USER,
+    ROLE_ADMIN
+}
