+++ src/main/java/inu/spp/cryptocoffee/auth/user/entity/UserEntity.java
@@ -1,0 +1,41 @@
+package inu.spp.cryptocoffee.auth.user.entity;
+
+import inu.spp.cryptocoffee.auth.user.dto.UserRoleEnum;
+import inu.spp.cryptocoffee.domain.entity.CompanyEntity;
+import inu.spp.cryptocoffee.global.base.BaseTimeEntity;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Email;
+import lombok.*;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
+@Getter
+@Table(name = "user")
+@Entity
+public class UserEntity extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long userId;
+
+    @Column(nullable = false, unique = true)
+    private String username;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private UserRoleEnum role;
+
+    @Column(nullable = false)
+    private final boolean isLocked = false;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private CompanyEntity company;
+}
