+++ src/main/java/inu/spp/cryptocoffee/auth/user/service/CustomUserDetailsService.java
@@ -1,0 +1,31 @@
+package inu.spp.cryptocoffee.auth.user.service;
+
+import inu.spp.cryptocoffee.auth.user.dto.CustomUserDetails;
+import inu.spp.cryptocoffee.auth.user.entity.UserEntity;
+import inu.spp.cryptocoffee.auth.user.repository.UserRepository;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Autowired
+    public CustomUserDetailsService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        UserEntity user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found with username: " + username));
+
+        return new CustomUserDetails(user);
+    }
+}
