+++ src/main/java/inu/spp/cryptocoffee/domain/entity/CompanyEntity.java
@@ -1,0 +1,28 @@
+package inu.spp.cryptocoffee.domain.entity;
+
+import inu.spp.cryptocoffee.domain.dto.Address;
+import inu.spp.cryptocoffee.global.base.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
+@Getter
+@Table(name = "company")
+public class CompanyEntity extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long companyId;
+
+    private String name;
+
+    @Column(name = "number")
+    private String companyNumber;
+
+    @Embedded
+    private Address address;
+}
