--- src/main/java/inu/spp/cryptocoffee/email/dto/EmailMessage.java
@@ -1,26 +1,0 @@
-package inu.spp.cryptocoffee.email.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class EmailMessage {
-    @Email
-    @NotNull(message = "email can't be null")
-    @NotEmpty(message = "email can't be empty")
-    private String to;
-
-    @NotNull(message = "email can't be null")
-    @NotEmpty(message = "email can't be empty")
-    private String subject;
-
-    @NotNull(message = "email can't be null")
-    @NotEmpty(message = "email can't be empty")
-    private String message;
-}
