--- src/main/java/inu/spp/cryptocoffee/user/controller/ReissueController.java
@@ -1,25 +1,0 @@
-package inu.spp.cryptocoffee.user.controller;
-
-import inu.spp.cryptocoffee.user.service.ReissueService;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Tag(name = "Token Reissue API")
-@RestController
-public class ReissueController {
-
-    private final ReissueService reissueService;
-
-    public ReissueController(ReissueService reissueService) {
-        this.reissueService = reissueService;
-    }
-
-    @PostMapping("/reissue")
-    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) {
-        return reissueService.reissueRefreshToken(request, response);
-    }
-}
