--- src/main/java/inu/spp/cryptocoffee/user/dto/CustomUserDetails.java
@@ -1,56 +1,0 @@
-package inu.spp.cryptocoffee.user.dto;
-
-import inu.spp.cryptocoffee.user.entity.UserEntity;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-public class CustomUserDetails implements UserDetails {
-
-    private final UserEntity user;
-
-    public CustomUserDetails(UserEntity user) {
-        this.user = user;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> collection = new ArrayList<>();
-
-        collection.add((GrantedAuthority) () -> user.getRole().toString());
-
-        return collection;
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return !user.isLocked();
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
