--- src/main/java/inu/spp/cryptocoffee/user/dto/UserRoleEnum.java
@@ -1,5 +1,0 @@
-package inu.spp.cryptocoffee.user.dto;
-
-public enum UserRoleEnum {
-    ROLE_USER, ROLE_MANAGER, ROLE_ADMIN
-}
