--- src/main/java/inu/spp/cryptocoffee/user/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package inu.spp.cryptocoffee.user.repository;
-
-import inu.spp.cryptocoffee.user.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-    Boolean existsByUsername(String username);
-
-    Optional<UserEntity> findByUsername(String username);
-
-}
